requirements.txt
annotated-types==0.7.0
anyio==4.4.0
build==1.0.3
CacheControl==0.14.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cleo==2.1.0
crashtest==0.4.1
cryptography==41.0.4
Cython==0.29.37
distlib==0.3.7
distro==1.9.0
dulwich==0.21.6
fastjsonschema==2.18.1
filelock==3.12.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
jaraco.classes==3.3.0
jeepney==0.8.0
keyring==24.2.0
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
meson==1.3.2
more-itertools==10.1.0
msgpack==1.0.7
openai==1.35.5
packaging==23.2
pexpect==4.9.0
pipenv==2024.0.0
pkginfo==1.10.0
platformdirs==3.11.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pycparser==2.21
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyproject_hooks==1.0.0
python-telegram-bot==21.3
rapidfuzz==3.5.2
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.3
six==1.16.0
sniffio==1.3.1
tomlkit==0.12.1
tqdm==4.66.4
trove-classifiers==2023.9.19
typing_extensions==4.12.2
urllib3==2.0.6
virtualenv==20.26.2
zipp==3.17.0
